Clarify difference between Struct and StructMap
The naming around Struct (a go struct wrapper) and the StructMap (a
map with a fixed set of keys) was confusing to say the least. This
commit cleans it up.
